+++ semantic_release/defaults.cfg
@@ -17,5 +17,5 @@ dist_path=dist
 remove_dist=true
 build_command=python setup.py sdist bdist_wheel
 branch=master
+changelog_components=semantic_release.changelog.changelog_headers
 changelog_sections=feature,fix,breaking,documentation,performance,:boom:,:sparkles:,:children_crossing:,:lipstick:,:iphone:,:egg:,:chart_with_upwards_trend:,:ambulance:,:lock:,:bug:,:zap:,:goal_net:,:alien:,:wheelchair:,:speech_balloon:,:mag:,:apple:,:penguin:,:checkered_flag:,:robot:,:green_apple:,Other
-compare_link=false
